+++ tests/test_cmdline.py
@@ -1,39 +1,72 @@
 from __future__ import annotations
 
+import os
+import tempfile
+from pathlib import Path
+
+import pytest
+import yaml
 from click.testing import CliRunner
-from pytest import TEST_DATA
 
 from duqtools import cli
+from duqtools.utils import work_directory
+
+imas = pytest.importorskip('imas',
+                           reason='No way of testing this without IMAS')
+
+CONTAINERIZED_RUNS_DIR = os.environ['CONTAINERIZED_RUNS_DIR']
+IMASDB = Path(CONTAINERIZED_RUNS_DIR).resolve() / 'imasdb'
+
+CONFIG = {
+    'tag': 'data_01',
+    'create': {
+        'runs_dir':
+        None,
+        'template_data': {
+            'user': str(IMASDB),
+            'db': 'jet',
+            'shot': 123,
+            'run': 1,
+        },
+        'operations': [
+            {
+                'variable': 't_e',
+                'operator': 'multiply',
+                'value': 0,
+            },
+        ],
+    },
+    'system': {
+        'name': 'nosystem',
+    },
+}
 
 
-def test_list_variables():
-    config = TEST_DATA / 'config_list-vars.yaml'
+@pytest.fixture(scope='module')
+def tmpworkdir():
+    jruns = os.environ.get('JRUNS', '.')
 
-    runner = CliRunner()
-    ret = runner.invoke(cli.cli_list_variables, [
-        '-c',
-        f'{config}',
-    ])
+    with tempfile.TemporaryDirectory(dir=jruns) as workdir:
+        CONFIG['create']['runs_dir'] = str(Path(workdir).resolve())
+        with open(Path(workdir) / 'duqtools.yaml', 'w') as f:
+            yaml.dump(CONFIG, f)
+        yield workdir
+
+
+@pytest.mark.dependency()
+def test_create(tmpworkdir):
+    with work_directory(tmpworkdir):
+        runner = CliRunner()
+        ret = runner.invoke(cli.cli_create, [
+            '--force',
+            '--yes',
+        ])
 
     assert ret.exit_code == 0
 
-    out = ret.output
-    assert 'IDS-variable' in out
-    assert 'jetto-variable' in out
-    assert '*my_extra_var' in out
-    assert 'rho_tor_norm' in out
+    outdir = Path(tmpworkdir)
+    for p in 'duqtools.log', 'runs.yaml', 'data.csv', 'run_0000':
+        assert (outdir / p).exists()
 
-
-def test_version():
-    runner = CliRunner()
-    ret = runner.invoke(cli.cli_version)
-
-    assert ret.exit_code == 0
-
-
-def test_no_command():
-    runner = CliRunner()
-    ret = runner.invoke(cli.cli)
-
-    assert ret.exit_code == 0
-    assert ret.output.startswith('Usage:')
+    for fn in 'ids_1230001.characteristics', 'ids_1230001.datafile', 'ids_1230001.tree':
+        assert Path(outdir, 'run_0000', 'imasdb', 'jet', '3', '0', fn).exists()
